\name{SBDN-package}
\alias{SBDN-package}
\alias{SBDN}
\docType{package}
\title{
Sparse Bayesian Dynamic Network model
}
\description{
Fits the Sparse Bayesian Dynamic Network model of Bartlett, Kosmidis and Silva (2020), implented as the function dynNetModelSamp.
}
\details{
\packageDESCRIPTION{SBDN}
\packageIndices{SBDN}
}
\author{
Thomas E. Bartlett, Ioannis Kosmidis, and Ricardo Silva.
}
\references{
Thomas E. Bartlett, Ioannis Kosmidis, and Ricardo Silva (2020). Two-way sparsity for time-varying networks, with applications in genomics. Preprint available: http://arxiv.org/abs/1802.08114
}
\keyword{}
\seealso{
}
\examples{
}
